+++ src/lib/sanity/image.ts
@@ -1,6 +1,7 @@
 import "server-only";
 
-import createImageUrlBuilder, {
+import {
+  createImageUrlBuilder,
   type SanityImageSource,
 } from "@sanity/image-url";